+++ govtool/frontend/src/consts/navItems.tsx
@@ -8,6 +8,7 @@ import {
   PATHS,
   PDF_PATHS,
   OUTCOMES_PATHS,
+  BUDGET_DISCUSSION_PATHS,
   // TODO: This will be uncommented when the page has been bootstrapped in the outcomes Pillar
   // USER_PATHS
 } from "./paths";
@@ -23,6 +24,12 @@ export const NAV_ITEMS = [
     dataTestId: "drep-directory-link",
     navTo: PATHS.dRepDirectory,
     label: i18n.t("dRepDirectory.title"),
+  },
+  {
+    dataTestId: "budget-discussion-link",
+    navTo: BUDGET_DISCUSSION_PATHS.budgetDiscussion,
+    label: i18n.t("budgetDiscussion.title"),
+    newTabLink: null,
   },
   {
     dataTestId: "proposed-governance-actions-link",
@@ -69,6 +76,14 @@ export const CONNECTED_NAV_ITEMS = [
     dataTestId: "drep-directory-link",
     label: i18n.t("dRepDirectory.title"),
     navTo: PATHS.dashboardDRepDirectory,
+    activeIcon: ICONS.dRepDirectoryActiveIcon,
+    icon: ICONS.dRepDirectoryIcon,
+    newTabLink: null,
+  },
+  {
+    dataTestId: "budget-discussion-link",
+    label: i18n.t("budgetDiscussion.title"),
+    navTo: BUDGET_DISCUSSION_PATHS.budgetDiscussion,
     activeIcon: ICONS.dRepDirectoryActiveIcon,
     icon: ICONS.dRepDirectoryIcon,
     newTabLink: null,